+++ test/runtests.jl
@@ -6,7 +6,9 @@ using S5
 @testset "S5.jl" begin
     include("test_ci_configuration.jl")
     include("test_governance.jl")
+    include("test_validation_policy.jl")
     include("test_utils.jl")
+    include("test_lrd_symbol_diagnostics.jl")
     include("test_pb1.jl")
     include("test_pb2.jl")
     include("test_pb3.jl")